+++ src/common.h
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#define CORNAMUSA_VERSION "0.8.1"
+#define CORNAMUSA_VERSION "0.8.2"
 #define CORNAMUSA_VERSION_MAJOR 0
 #define CORNAMUSA_VERSION_MINOR 8
-#define CORNAMUSA_VERSION_PATCH 1
+#define CORNAMUSA_VERSION_PATCH 2
 
 #endif /* CORNAMUSA_COMMON_H */